requirements.txt
Flask==2.2.2
marshmallow==3.17.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0